Setting up render workflow
