+++ .env.example
@@ -73,6 +73,12 @@ NODE_ENV=development
 # Generate with: openssl rand -base64 32
 # FTS_API_SECRET=your-shared-secret-here
 
+# ── Task Queue (Render Workflows) ──────────────────────────────────────
+# Set both to enable delegating long-running operations (sync, ingest)
+# to a Render Workflow service. Required for sync on serverless runtimes.
+# RENDER_WORKFLOW_SLUG=
+# RENDER_API_KEY=
+
 # ── Runtime Override ────────────────────────────────────────────────────────
 # Auto-detected from platform env vars. Set this to override detection.
 # Options: vercel | aws_lambda | netlify | fly | railway | render | docker | development
