+++ apps/web/server/jobs/dispatch.ts
@@ -1,0 +1,6 @@
+import { runtime } from "../runtime-context";
+
+export function shouldDelegateToWorkflow(): boolean {
+  if (runtime.longRunningSupported) return false;
+  return !!process.env.RENDER_WORKFLOW_SLUG;
+}
