+++ apps/web/server/s3/auth.ts
@@ -1,42 +1,16 @@
-import { createHmac, createHash, randomBytes, createCipheriv, createDecipheriv } from 'crypto';
+import { createHmac, createHash, randomBytes } from 'crypto';
 import { getDb } from '@locker/database/client';
 import { s3ApiKeys, workspaces } from '@locker/database';
 import { eq, and } from 'drizzle-orm';
 
+// Re-export from shared package — crypto is used by both web app and workflow service
+export { encryptSecret, decryptSecret } from '@locker/jobs';
+// Local import for SigV4 verification below
+import { decryptSecret } from '@locker/jobs';
+
 const ACCESS_KEY_PREFIX = 'LKAK';
 const ACCESS_KEY_LENGTH = 20;
 const SECRET_KEY_BYTES = 40;
-const ALGORITHM = 'aes-256-gcm';
-
-// ── Key encryption/decryption ───────────────────────────────────────────
-
-function getEncryptionKey(): Buffer {
-  const secret = process.env.S3_API_KEY_ENCRYPTION_SECRET;
-  if (!secret) throw new Error('S3_API_KEY_ENCRYPTION_SECRET not configured');
-  return createHash('sha256').update(secret).digest();
-}
-
-export function encryptSecret(plaintext: string): string {
-  const key = getEncryptionKey();
-  const iv = randomBytes(16);
-  const cipher = createCipheriv(ALGORITHM, key, iv);
-  const encrypted = Buffer.concat([cipher.update(plaintext, 'utf8'), cipher.final()]);
-  const tag = cipher.getAuthTag();
-  // Format: iv:tag:ciphertext (all base64)
-  return `${iv.toString('base64')}:${tag.toString('base64')}:${encrypted.toString('base64')}`;
-}
-
-export function decryptSecret(stored: string): string {
-  const key = getEncryptionKey();
-  const [ivB64, tagB64, dataB64] = stored.split(':');
-  if (!ivB64 || !tagB64 || !dataB64) throw new Error('Invalid encrypted secret format');
-  const iv = Buffer.from(ivB64, 'base64');
-  const tag = Buffer.from(tagB64, 'base64');
-  const data = Buffer.from(dataB64, 'base64');
-  const decipher = createDecipheriv(ALGORITHM, key, iv);
-  decipher.setAuthTag(tag);
-  return decipher.update(data) + decipher.final('utf8');
-}
 
 // ── Key generation ──────────────────────────────────────────────────────
