+++ apps/web/server/trpc/routers/stores.ts
@@ -9,9 +9,11 @@ import {
 } from "@locker/database";
 import { createRouter, workspaceAdminProcedure } from "../init";
 import { createStorageFromConfig, type WorkspaceStorageConfig } from "@locker/storage";
-import { getStoreById, saveStoreSecret } from "../../storage";
+import { getStoreById, saveStoreSecret, makeWebFileSourceResolver } from "../../storage";
 import { syncWorkspaceStores, ingestFromReadOnlyStore } from "../../stores/sync";
 import { runtime } from "../../runtime-context";
+import { shouldDelegateToWorkflow } from "../../jobs/dispatch";
+import { dispatchSyncWorkspace } from "../../jobs/workflow-client";
 
 const providerSchema = z.enum(["s3", "r2", "vercel_blob", "local"]);
 const writeModeSchema = z.enum(["write", "read_only"]);
@@ -508,17 +510,59 @@ export const storesRouter = createRouter({
         .optional(),
     )
     .mutation(async ({ ctx, input }) => {
-      if (!runtime.longRunningSupported) {
+      if (!runtime.longRunningSupported && !runtime.taskQueueAvailable) {
         throw new TRPCError({
           code: "PRECONDITION_FAILED",
-          message: `Store sync requires a persistent runtime. This operation is not supported on ${runtime.environment}.`,
+          message: `Store sync requires a persistent runtime or task queue. Not supported on ${runtime.environment}.`,
         });
+      }
+
+      if (shouldDelegateToWorkflow()) {
+        const [run] = await ctx.db
+          .insert(replicationRuns)
+          .values({
+            workspaceId: ctx.workspaceId,
+            kind: "manual_sync",
+            status: "queued",
+            targetStoreId: input?.storeId ?? null,
+            triggeredByUserId: ctx.userId,
+          })
+          .returning({ id: replicationRuns.id });
+
+        try {
+          await dispatchSyncWorkspace({
+            runId: run!.id,
+            workspaceId: ctx.workspaceId,
+            targetStoreId: input?.storeId,
+            triggeredByUserId: ctx.userId,
+          });
+        } catch (err) {
+          await ctx.db
+            .update(replicationRuns)
+            .set({
+              status: "failed",
+              errorMessage:
+                err instanceof Error
+                  ? err.message
+                  : "Failed to dispatch to task queue",
+              completedAt: new Date(),
+              updatedAt: new Date(),
+            })
+            .where(eq(replicationRuns.id, run!.id));
+          throw new TRPCError({
+            code: "INTERNAL_SERVER_ERROR",
+            message: "Failed to dispatch sync to task queue",
+          });
+        }
+
+        return { runId: run!.id };
       }
 
       return syncWorkspaceStores({
         workspaceId: ctx.workspaceId,
         targetStoreId: input?.storeId,
         triggeredByUserId: ctx.userId,
+        resolveFileSource: makeWebFileSourceResolver(),
       });
     }),
