+++ apps/workflows/src/index.ts
@@ -1,0 +1,1 @@
+import "./tasks/sync-workspace";
