+++ apps/workflows/src/tasks/sync-workspace.ts
@@ -1,0 +1,25 @@
+import { task } from "@renderinc/sdk/workflows";
+import { syncWorkspaceStores } from "@locker/jobs";
+import { resolveFileSource } from "../resolve-file-source";
+
+export const syncWorkspace = task(
+  {
+    name: "syncWorkspace",
+    timeoutSeconds: 3600,
+    retry: { maxRetries: 1, waitDurationMs: 5000 },
+  },
+  async function syncWorkspace(
+    runId: string,
+    workspaceId: string,
+    targetStoreId: string | null,
+    triggeredByUserId: string | null,
+  ) {
+    return syncWorkspaceStores({
+      runId,
+      workspaceId,
+      targetStoreId: targetStoreId ?? undefined,
+      triggeredByUserId: triggeredByUserId ?? undefined,
+      resolveFileSource,
+    });
+  },
+);
